+++ src/modules/auth/AuthContext.tsx
@@ -1,6 +1,11 @@
 import { createContext } from 'react';
 
-import { HmisAuthState } from '../hmisAppSettings/types';
+import { HmisUser } from '@/modules/auth/api/sessions';
+
+export interface HmisAuthState {
+  user?: HmisUser;
+  setUser: (user: HmisUser | undefined) => void;
+}
 
 const initialState: HmisAuthState = {
   user: undefined,
